feat: move/rename pages and tests
