+++ src/app/(registration)/onboarding/role-selection/role-selection.test.tsx
@@ -1,7 +1,8 @@
 import { screen } from "@testing-library/dom";
 
+import RoleSelectionPage from "./page";
+
 import renderWithProviders from "@/__tests__/__helpers__/renderWithProviders";
-import RoleSelectionPage from "@/pages/onboarding/role-selection";
 import { mockLoggedIn } from "@/__tests__/__helpers__/mockUser";
 import { setUseUserReturn } from "@/__tests__/__helpers__/mockClerk";
