+++ src/app/(registration)/onboarding/school-selection/page.tsx
